+++ README.md
@@ -1,2 +1,3 @@
 JSAutomation
 3. Домашние электроприборы. Определить иерархию электроприборов. Включить некоторые в розетку. Посчитать потребляемую мощность. Провести сортировку приборов в квартире на основе одного из параметров. Найти прибор в квартире, соответствующий заданному диапазону параметров.
+//readline-sync
