+++ devices/mainDevice.js
@@ -4,8 +4,8 @@ class MainDevice {
         this.room = room;
         this.power = power;
         this.weight = weight;
-        this.isPowerOn = isPowerOn !== 'false' ? true : false;
-        this.isTurnedOn = isTurnedOn !== 'false' ? true : false;
+        this.isPowerOn = isPowerOn; //!== 'false' ? true : false;
+        this.isTurnedOn = isTurnedOn; //!== 'false' ? true : false;
     }
 };
 module.exports = MainDevice;
